swagger

+++ src/main/java/com/netposa/lucai/util/ImgUtils.java
@@ -17,7 +17,7 @@ public class ImgUtils {
 
 	public static String tempPath;
 
-	public static Pair<String,byte[]> writeImg(MultipartFile file) {
+	public static Pair<String, byte[]> writeImg(MultipartFile file) {
 		String fileName;
 		InputStream in = null;
 		OutputStream out = null;
@@ -26,24 +26,28 @@ public class ImgUtils {
 			out = new FileOutputStream(getFilePathTemp(fileName));
 			in = file.getInputStream();
 			IOUtils.copy(in, out);
-			return Pair.of(fileName,file.getBytes());
+			return Pair.of(fileName, file.getBytes());
 		} catch (Exception e) {
 			throw new BusinessException("上传的文件不正确", e);
-		}finally {
-			IOUtils.closeQuietly(in,out);
+		} finally {
+			IOUtils.closeQuietly(in, out);
 		}
 	}
 
+	//文件全路径
 	private static String getFilePath(String prefix, String fileName) {
 		return basePath + File.separator + prefix + File.separator + fileName;
 	}
 
+	//归档目录
 	private static String getArchiveDir(String prefix) {
 		return basePath + File.separator + prefix;
 	}
 
+
+	//临时目录
 	private static String getFilePathTemp(String fileName) {
-		return basePath + File.separator + tempPath + File.separator + fileName ;
+		return basePath + File.separator + tempPath + File.separator + fileName;
 	}
 
 	@SuppressWarnings("all")
@@ -54,12 +58,23 @@ public class ImgUtils {
 
 	public static void archiveFile(String fileName, String prefix) {
 		try {
-			File srcfile = new File(getFilePathTemp(fileName));
-			File descfile = new File(getArchiveDir(prefix));
-			FileUtils.moveFileToDirectory(srcfile, descfile, true);
+			if(!isExists(fileName,prefix)){//文件已经存在就不归档了
+				File srcfile = new File(getFilePathTemp(fileName));
+				File descfile = new File(getArchiveDir(prefix));
+				FileUtils.moveFileToDirectory(srcfile, descfile, true);
+			}
 		} catch (IOException e) {
-			log.error(String.format("文件归档失败:%s",e.getMessage()),e);
+			log.error(String.format("文件归档失败:%s", e.getMessage()), e);
 		}
 	}
 
+
+	public static boolean isExists(String fileName, String prefix) {
+		return new File(basePath + File.separator
+				+ prefix + File.separator + fileName).exists();
+	}
+
+	public static void delImgByCamera(String id) {
+		new File(getArchiveDir(id)).delete();
+	}
 }
